+++ test/integration/distributed/backup/backup_suite_test.go
@@ -15,7 +15,7 @@
 // specific language governing permissions and limitations
 // under the License.
 
-package snapshot_test
+package backup_test
 
 import (
 	"context"
@@ -48,12 +48,12 @@ import (
 	test_stream "github.com/apache/skywalking-banyandb/pkg/test/stream"
 	"github.com/apache/skywalking-banyandb/pkg/timestamp"
 	test_cases "github.com/apache/skywalking-banyandb/test/cases"
-	casessnapshot "github.com/apache/skywalking-banyandb/test/cases/snapshot"
+	casesbackup "github.com/apache/skywalking-banyandb/test/cases/backup"
 )
 
-func TestSnapshot(t *testing.T) {
+func TestBackup(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Distributed Snapshot Suite")
+	RunSpecs(t, "Distributed Backup Suite")
 }
 
 var (
@@ -61,6 +61,7 @@ var (
 	dir        string
 	deferFunc  func()
 	goods      []gleak.Goroutine
+	dataAddr   string
 )
 
 var _ = SynchronizedBeforeSuite(func() []byte {
@@ -92,7 +93,6 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	test_stream.PreloadSchema(ctx, schemaRegistry)
 	test_measure.PreloadSchema(ctx, schemaRegistry)
 	By("Starting data node 0")
-	var dataAddr string
 	var closeDataNode0 func()
 	dataAddr, dir, closeDataNode0 = setup.DataNodeWithAddrAndDir(ep)
 	By("Starting liaison node")
@@ -146,7 +146,8 @@ var _ = SynchronizedBeforeSuite(func() []byte {
 	connection, err = grpchelper.Conn(string(address), 10*time.Second,
 		grpc.WithTransportCredentials(insecure.NewCredentials()))
 	Expect(err).NotTo(HaveOccurred())
-	casessnapshot.SharedContext = helpers.SnapshotSharedContext{
+	casesbackup.SharedContext = helpers.BackupSharedContext{
+		DataAddr:   dataAddr,
 		Connection: connection,
 		RootDir:    dir,
 	}